Move connectToAllConfiguredChains into Persona
